Backend/requirements.txt
fastapi
uvicorn
spacy
python-multipart
jinja2
google
python-dotenv
pydantic
mangum